refactor(taxes): extract ChecklistRow component
Move ChecklistRow to src/pages/taxes/components/ChecklistRow.tsx.
Taxes.tsx 995→859L.

Part of #181
